# docs/api/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoEDA API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/datasets/upload": {
      "post": {
        "summary": "Datasets Upload",
        "operationId": "datasets_upload_api_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_datasets_upload_api_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/eda": {
      "post": {
        "summary": "Eda",
        "operationId": "eda_api_eda_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EDARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EDAReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/charts/suggest": {
      "post": {
        "summary": "Charts Suggest",
        "operationId": "charts_suggest_api_charts_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartsSuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartsSuggestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/qna": {
      "post": {
        "summary": "Qna",
        "operationId": "qna_api_qna_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QnARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QnAResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/actions/prioritize": {
      "post": {
        "summary": "Prioritize",
        "operationId": "prioritize_api_actions_prioritize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrioritizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrioritizedAction"
                  },
                  "type": "array",
                  "title": "Response Prioritize Api Actions Prioritize Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pii/scan": {
      "post": {
        "summary": "Pii Scan",
        "operationId": "pii_scan_api_pii_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PIIScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PIIScanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pii/apply": {
      "post": {
        "summary": "Pii Apply",
        "operationId": "pii_apply_api_pii_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PIIApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PIIApplyResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/leakage/scan": {
      "post": {
        "summary": "Leakage Scan",
        "operationId": "leakage_scan_api_leakage_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeakageScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeakageScanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/leakage/resolve": {
      "post": {
        "summary": "Leakage Resolve",
        "operationId": "leakage_resolve_api_leakage_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeakageResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeakageScanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/emit": {
      "post": {
        "summary": "Recipes Emit",
        "operationId": "recipes_emit_api_recipes_emit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeEmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeEmitResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/export": {
      "post": {
        "summary": "Recipes Export",
        "operationId": "recipes_export_api_recipes_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeEmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeEmitResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_datasets_upload_api_datasets_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_datasets_upload_api_datasets_upload_post"
      },
...
      "ChartsSuggestRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "k": {
            "type": "integer",
            "title": "K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "ChartsSuggestRequest"
      },
      "ChartsSuggestResponse": {
        "properties": {
          "charts": {
            "items": {
              "$ref": "#/components/schemas/ChartCandidate"
            },
            "type": "array",
            "title": "Charts"
          }
        },
        "type": "object",
        "required": [
          "charts"
        ],
        "title": "ChartsSuggestResponse"
      },
...
      "EDAReport": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "distributions": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distributions"
          },
          "key_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Features"
          },
          "outliers": {
            "items": {
              "$ref": "#/components/schemas/Outlier"
            },
            "type": "array",
            "title": "Outliers"
          },
          "data_quality_report": {
            "$ref": "#/components/schemas/DataQualityReport"
          },
          "next_actions": {
            "items": {
              "$ref": "#/components/schemas/NextAction"
            },
            "type": "array",
            "title": "Next Actions"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/Reference"
            },
            "type": "array",
            "title": "References"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "distributions",
          "key_features",
          "outliers",
          "data_quality_report",
          "next_actions",
          "references"
        ],
        "title": "EDAReport"
      },
      "EDARequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "sample_ratio": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Ratio"
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "EDARequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LeakageResolveRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "exclude",
              "acknowledge",
              "reset"
            ],
            "title": "Action",
            "default": "exclude"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "columns"
        ],
        "title": "LeakageResolveRequest"
      },
      "LeakageScanRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "LeakageScanRequest"
      },
      "LeakageScanResult": {
        "properties": {
          "flagged_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flagged Columns"
          },
          "rules_matched": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules Matched"
          },
          "excluded_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Columns",
            "default": []
          },
          "acknowledged_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Acknowledged Columns",
            "default": []
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "flagged_columns",
          "rules_matched"
        ],
        "title": "LeakageScanResult"
      },
...
      "PIIApplyRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "mask_policy": {
            "type": "string",
            "enum": [
              "MASK",
              "HASH",
              "DROP"
            ],
            "title": "Mask Policy",
            "default": "MASK"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "PIIApplyRequest"
      },
      "PIIApplyResult": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "mask_policy": {
            "type": "string",
            "enum": [
              "MASK",
              "HASH",
              "DROP"
            ],
            "title": "Mask Policy"
          },
          "masked_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Masked Fields"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "mask_policy",
          "masked_fields",
          "updated_at"
        ],
        "title": "PIIApplyResult"
      },
      "PIIScanRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "PIIScanRequest"
      },
      "PIIScanResult": {
        "properties": {
          "detected_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detected Fields"
          },
          "mask_policy": {
            "type": "string",
            "enum": [
              "MASK",
              "HASH",
              "DROP"
            ],
            "title": "Mask Policy",
            "default": "MASK"
          },
          "masked_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Masked Fields",
            "default": []
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "detected_fields"
        ],
        "title": "PIIScanResult"
      },
      "PrioritizeRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "next_actions": {
            "items": {
              "$ref": "#/components/schemas/PrioritizeRequestItem"
            },
            "type": "array",
            "title": "Next Actions"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "next_actions"
        ],
        "title": "PrioritizeRequest"
      },
...
      "PrioritizedAction": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "impact": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Impact"
          },
          "effort": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Effort"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "wsjf": {
            "type": "number",
            "title": "Wsjf"
          },
          "rice": {
            "type": "number",
            "title": "Rice"
          },
          "dependencies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencies"
          }
        },
        "type": "object",
        "required": [
          "title",
          "impact",
          "effort",
          "confidence",
          "score",
          "wsjf",
          "rice"
        ],
        "title": "PrioritizedAction"
      },
      "QnARequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "question"
        ],
        "title": "QnARequest"
      },
      "QnAResponse": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/Answer"
            },
            "type": "array",
            "title": "Answers"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/Reference"
            },
            "type": "array",
            "title": "References",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "QnAResponse"
      },
      "RecipeEmitRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "RecipeEmitRequest"
      },
      "RecipeEmitResult": {
        "properties": {
          "artifact_hash": {
            "type": "string",
            "title": "Artifact Hash"
          },
          "files": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Files"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Summary"
              },
              {
                "type": "null"
              }
            ]
          },
          "measured_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured Summary"
          }
        },
        "type": "object",
        "required": [
          "artifact_hash",
          "files"
        ],
        "title": "RecipeEmitResult"
      },
...
      "UploadResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          }
        },
        "type": "object",
        "required": [
          "dataset_id"
        ],
        "title": "UploadResponse"
      },
...
    }
  }
}